+++ Cargo.toml
@@ -24,7 +24,7 @@ serde_json = "1.0.140"
 
 strsim = "0.11.1"
 
-tokio = { version = "1.45.0", features = ["full"] }
+tokio = { version = "1.45.0", features = ["macros", "rt-multi-thread"] }
 single-instance = "0.3.3"
 md5 = "0.7.0"
 env_logger = "0.11.8"
