--- src/pages/popups/link.rs
@@ -1,2 +1,0 @@
-pub mod config;
-pub mod save;
